Revert "Merge pull request #703 from mirumee/ref/sidebar-mobile"
This reverts commit acfd04e81af1d4f21de3331c26acfd935aa34870, reversing
changes made to 40d27a92805db9cd2b4e7029571e670b1196d892.
