+++ src/components/AppLayout/AppLayout.tsx
@@ -1,14 +1,23 @@
+import Avatar from "@material-ui/core/Avatar";
+import Chip from "@material-ui/core/Chip";
+import ClickAwayListener from "@material-ui/core/ClickAwayListener";
+import Grow from "@material-ui/core/Grow";
 import LinearProgress from "@material-ui/core/LinearProgress";
-import { makeStyles, Theme } from "@material-ui/core/styles";
-import useMediaQuery from "@material-ui/core/useMediaQuery";
+import MenuItem from "@material-ui/core/MenuItem";
+import Menu from "@material-ui/core/MenuList";
+import Paper from "@material-ui/core/Paper";
+import Popper from "@material-ui/core/Popper";
+import { makeStyles } from "@material-ui/core/styles";
 import { createConfigurationMenu } from "@saleor/configuration";
 import useAppState from "@saleor/hooks/useAppState";
 import useNavigator from "@saleor/hooks/useNavigator";
 import useTheme from "@saleor/hooks/useTheme";
 import useUser from "@saleor/hooks/useUser";
+import ArrowDropdown from "@saleor/icons/ArrowDropdown";
 import { staffMemberDetailsUrl } from "@saleor/staff/urls";
+import classNames from "classnames";
 import React from "react";
-import { useIntl } from "react-intl";
+import { FormattedMessage, useIntl } from "react-intl";
 import useRouter from "use-react-router";
 
 import Container from "../Container";
@@ -16,8 +25,6 @@ import ErrorPage from "../ErrorPage";
 import Navigator from "../Navigator";
 import NavigatorButton from "../NavigatorButton/NavigatorButton";
 import SideBar from "../SideBar";
-import SideBarDrawer from "../SideBarDrawer/SideBarDrawer";
-import UserChip from "../UserChip";
 import AppActionContext from "./AppActionContext";
 import AppHeaderContext from "./AppHeaderContext";
 import { appLoaderHeight } from "./consts";
@@ -45,13 +52,22 @@ const useStyles = makeStyles(
       height: appLoaderHeight,
       marginBottom: theme.spacing(4)
     },
-
+    arrow: {
+      marginLeft: theme.spacing(2),
+      transition: theme.transitions.duration.standard + "ms"
+    },
+    avatar: {
+      "&&": {
+        height: 32,
+        width: 32
+      }
+    },
     content: {
       flex: 1
     },
     darkThemeSwitch: {
       [theme.breakpoints.down("sm")]: {
-        marginRight: theme.spacing(1)
+        marginRight: -theme.spacing(1.5)
       },
       marginRight: theme.spacing(2)
     },
@@ -59,8 +75,8 @@ const useStyles = makeStyles(
       display: "grid",
       gridTemplateAreas: `"headerAnchor headerToolbar"`,
       [theme.breakpoints.down("sm")]: {
-        gridTemplateAreas: `"headerToolbar" 
-        "headerAnchor"`
+        height: 88,
+        marginBottom: 0
       },
       marginBottom: theme.spacing(3)
     },
@@ -75,20 +91,54 @@ const useStyles = makeStyles(
         height: "auto"
       }
     },
+    menu: {
+      background: theme.palette.background.paper,
+      height: "100vh",
+      padding: "25px 20px"
+    },
+    menuSmall: {
+      background: theme.palette.background.paper,
+      height: "100vh",
+      overflow: "hidden",
+      padding: 25
+    },
+    popover: {
+      zIndex: 1
+    },
     root: {
       [theme.breakpoints.up("md")]: {
         display: "flex"
       },
       width: `100%`
     },
+    rotate: {
+      transform: "rotate(180deg)"
+    },
     spacer: {
       flex: 1
     },
     userBar: {
+      [theme.breakpoints.down("sm")]: {
+        alignItems: "flex-end",
+        flexDirection: "column-reverse",
+        overflow: "hidden"
+      },
       alignItems: "center",
       display: "flex"
     },
-
+    userChip: {
+      backgroundColor: theme.palette.background.paper,
+      borderRadius: 24,
+      color: theme.palette.text.primary,
+      height: 40,
+      padding: theme.spacing(0.5)
+    },
+    userMenuContainer: {
+      position: "relative"
+    },
+    userMenuItem: {
+      textAlign: "right"
+    },
     view: {
       flex: 1,
       flexGrow: 1,
@@ -114,15 +164,16 @@ interface AppLayoutProps {
 const AppLayout: React.FC<AppLayoutProps> = ({ children }) => {
   const classes = useStyles({});
   const { isDark, toggleTheme } = useTheme();
+  const [isMenuOpened, setMenuState] = React.useState(false);
   const appActionAnchor = React.useRef<HTMLDivElement>();
   const appHeaderAnchor = React.useRef<HTMLDivElement>();
+  const anchor = React.useRef<HTMLDivElement>();
   const { logout, user } = useUser();
   const navigate = useNavigator();
   const intl = useIntl();
   const [appState, dispatchAppState] = useAppState();
   const { location } = useRouter();
   const [isNavigatorVisible, setNavigatorVisibility] = React.useState(false);
-  const isMdUp = useMediaQuery((theme: Theme) => theme.breakpoints.up("md"));
 
   const menuStructure = createMenuStructure(intl);
   const configurationMenu = createConfigurationMenu(intl);
@@ -136,6 +187,16 @@ const AppLayout: React.FC<AppLayoutProps> = ({ children }) => {
         )
     )
   );
+
+  const handleLogout = () => {
+    setMenuState(false);
+    logout();
+  };
+
+  const handleViewerProfile = () => {
+    setMenuState(false);
+    navigate(staffMemberDetailsUrl(user.id));
+  };
 
   const handleErrorBack = () => {
     navigate("/");
@@ -156,15 +217,13 @@ const AppLayout: React.FC<AppLayoutProps> = ({ children }) => {
       <AppHeaderContext.Provider value={appHeaderAnchor}>
         <AppActionContext.Provider value={appActionAnchor}>
           <div className={classes.root}>
-            {isMdUp && (
-              <SideBar
-                menuItems={menuStructure}
-                location={location.pathname}
-                user={user}
-                renderConfigure={renderConfigure}
-                onMenuItemClick={navigate}
-              />
-            )}
+            <SideBar
+              menuItems={menuStructure}
+              location={location.pathname}
+              user={user}
+              renderConfigure={renderConfigure}
+              onMenuItemClick={navigate}
+            />
             <div className={classes.content}>
               {appState.loading ? (
                 <LinearProgress className={classes.appLoader} color="primary" />
@@ -175,40 +234,93 @@ const AppLayout: React.FC<AppLayoutProps> = ({ children }) => {
                 <div>
                   <Container>
                     <div className={classes.header}>
-                      <div
-                        className={classes.headerAnchor}
-                        ref={appHeaderAnchor}
-                      />
-                      <div className={classes.headerToolbar}>
-                        {!isMdUp && (
-                          <SideBarDrawer
-                            menuItems={menuStructure}
-                            location={location.pathname}
-                            user={user}
-                            renderConfigure={renderConfigure}
-                            onMenuItemClick={navigate}
-                          />
-                        )}
-                        <div className={classes.spacer} />
-                        <div className={classes.userBar}>
-                          <ThemeSwitch
-                            className={classes.darkThemeSwitch}
-                            checked={isDark}
-                            onClick={toggleTheme}
-                          />
-                          <NavigatorButton
-                            isMac={navigator.platform
-                              .toLowerCase()
-                              .includes("mac")}
-                            onClick={() => setNavigatorVisibility(true)}
-                          />
-                          <UserChip
-                            onLogout={logout}
-                            onProfileClick={() =>
-                              navigate(staffMemberDetailsUrl(user.id))
+                      <div ref={appHeaderAnchor} />
+                      <div className={classes.spacer} />
+                      <div className={classes.userBar}>
+                        <ThemeSwitch
+                          className={classes.darkThemeSwitch}
+                          checked={isDark}
+                          onClick={toggleTheme}
+                        />
+                        <NavigatorButton
+                          isMac={navigator.platform
+                            .toLowerCase()
+                            .includes("mac")}
+                          onClick={() => setNavigatorVisibility(true)}
+                        />
+                        <div className={classes.userMenuContainer} ref={anchor}>
+                          <Chip
+                            avatar={
+                              user.avatar && (
+                                <Avatar alt="user" src={user.avatar.url} />
+                              )
                             }
-                            user={user}
+                            classes={{
+                              avatar: classes.avatar
+                            }}
+                            className={classes.userChip}
+                            label={
+                              <>
+                                {user.email}
+                                <ArrowDropdown
+                                  className={classNames(classes.arrow, {
+                                    [classes.rotate]: isMenuOpened
+                                  })}
+                                />
+                              </>
+                            }
+                            onClick={() => setMenuState(!isMenuOpened)}
+                            data-test="userMenu"
                           />
+                          <Popper
+                            className={classes.popover}
+                            open={isMenuOpened}
+                            anchorEl={anchor.current}
+                            transition
+                            placement="bottom-end"
+                          >
+                            {({ TransitionProps, placement }) => (
+                              <Grow
+                                {...TransitionProps}
+                                style={{
+                                  transformOrigin:
+                                    placement === "bottom"
+                                      ? "right top"
+                                      : "right bottom"
+                                }}
+                              >
+                                <Paper>
+                                  <ClickAwayListener
+                                    onClickAway={() => setMenuState(false)}
+                                    mouseEvent="onClick"
+                                  >
+                                    <Menu>
+                                      <MenuItem
+                                        className={classes.userMenuItem}
+                                        onClick={handleViewerProfile}
+                                        data-test="accountSettingsButton"
+                                      >
+                                        <FormattedMessage
+                                          defaultMessage="Account Settings"
+                                          description="button"
+                                        />
+                                      </MenuItem>
+                                      <MenuItem
+                                        className={classes.userMenuItem}
+                                        onClick={handleLogout}
+                                        data-test="logOutButton"
+                                      >
+                                        <FormattedMessage
+                                          defaultMessage="Log out"
+                                          description="button"
+                                        />
+                                      </MenuItem>
+                                    </Menu>
+                                  </ClickAwayListener>
+                                </Paper>
+                              </Grow>
+                            )}
+                          </Popper>
                         </div>
                       </div>
                     </div>
