+++ src/components/SideBar/SideBar.tsx
@@ -9,7 +9,7 @@ import { sectionNames } from "@saleor/intl";
 import classNames from "classnames";
 import React from "react";
 import SVG from "react-inlinesvg";
-import { IntlShape, useIntl } from "react-intl";
+import { useIntl } from "react-intl";
 
 import { IMenuItem } from "../AppLayout/menuStructure";
 import ExpandButton from "./ExpandButton";
@@ -55,14 +55,6 @@ export interface IActiveSubMenu {
   label: string | null;
 }
 
-export const getConfigureMenuItem = (intl: IntlShape): IMenuItem => ({
-  ariaLabel: "configure",
-  icon: configurationIcon,
-  label: intl.formatMessage(sectionNames.configuration),
-  testingContextId: "configure",
-  url: configurationMenuUrl
-});
-
 const SideBar: React.FC<SideBarProps> = ({
   location,
   menuItems,
@@ -73,7 +65,6 @@ const SideBar: React.FC<SideBarProps> = ({
   const classes = useStyles({});
   const [isShrunk, setShrink] = useLocalStorage("isMenuSmall", false);
   const intl = useIntl();
-  const configureMenuItem = getConfigureMenuItem(intl);
 
   return (
     <div
@@ -116,7 +107,13 @@ const SideBar: React.FC<SideBarProps> = ({
               )
             }
             isMenuShrunk={isShrunk}
-            menuItem={configureMenuItem}
+            menuItem={{
+              ariaLabel: "configure",
+              icon: configurationIcon,
+              label: intl.formatMessage(sectionNames.configuration),
+              testingContextId: "configure",
+              url: configurationMenuUrl
+            }}
             onClick={onMenuItemClick}
           />
         )}
