--- src/components/SquareButton/index.ts
@@ -1,2 +1,0 @@
-export * from "./SquareButton";
-export { default } from "./SquareButton";
