--- src/components/UserChip/index.ts
@@ -1,2 +1,0 @@
-export * from "./UserChip";
-export { default } from "./UserChip";
